requirements.txt
tensorflow
numpy
Pillow
opencv-python
matplotlib
tensorflow_hub
pandas
gradio